+++ src/main/java/br/edu/ifpb/acomidadobebeservice/controller/LoginController.java
@@ -1,14 +1,13 @@
 package br.edu.ifpb.acomidadobebeservice.controller;
 
 import java.nio.charset.StandardCharsets;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,6 +18,9 @@ import br.edu.ifpb.acomidadobebeservice.model.Login;
 import br.edu.ifpb.acomidadobebeservice.model.Usuario;
 import br.edu.ifpb.acomidadobebeservice.repository.LoginRepository;
 import br.edu.ifpb.acomidadobebeservice.repository.UsuarioRepository;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+
 
 @RestController
 public class LoginController {
@@ -29,76 +31,86 @@ public class LoginController {
     private UsuarioRepository _usuarioRepository;
 
     // Listar todos
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public List<Login> get() {
+    @RequestMapping(value = "/loginTeste", method = RequestMethod.GET)
+    public List<Login> Get() {
         return _loginRepository.findAll();
     }
 
     // Listar pelo id
-    @RequestMapping(value = "/login/{id}", method = RequestMethod.GET)
+    @RequestMapping(value = "/loginTeste/{id}", method = RequestMethod.GET)
     public ResponseEntity<Login> getById(@PathVariable(value = "id") Integer id)
     {
-        Optional<Login> login = _loginRepository.findById(id);
-        if(login.isPresent())
-            return new ResponseEntity<Login>(login.get(), HttpStatus.OK);
+        Optional<Login> loginTeste = _loginRepository.findById(id);
+        if(loginTeste.isPresent())
+            return new ResponseEntity<Login>(loginTeste.get(), HttpStatus.OK);
         else
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Listar pelo token
-    @RequestMapping(value = "/login/token/{token}", method = RequestMethod.GET)
-    public ResponseEntity<Login> getByToken(@PathVariable(value = "token") Integer token)
-    {
-        Optional<Login> login = _loginRepository.findByToken(token);
-        if(login.isPresent())
-            return new ResponseEntity<Login>(login.get(), HttpStatus.OK);
-        else
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-    }
+    /* 
+    * TODO Lista pelo token
+    */ 
+
+    // Lista usuario pelo token
+    //@RequestMapping(value = "/loginTeste/token/{token}", method = RequestMethod.GET)
+    //public List<Usuario> getByToken(@PathVariable(value = "token") String token) {
+    //   return _usuarioRepository.findByLoginTesteToken(token);
+    //}
 
     // Cadastrar
-    @RequestMapping(value = "/login", method =  RequestMethod.POST)
-    public Login post(@RequestBody Login login)
+    @RequestMapping(value = "/loginTeste", method =  RequestMethod.POST)
+    public ResponseEntity<Login> Post(@RequestBody Login loginTeste)
     {
-        String token = Hashing.sha256()
-            .hashString(login.getSenha(), StandardCharsets.UTF_8)
-            .toString();
+        Optional<Usuario> usuarioOptional = _usuarioRepository.findById(loginTeste.getUsuario().getId()); // mudar para optional
+        if(usuarioOptional.isPresent()){
+            Usuario usuario = usuarioOptional.get();
+            loginTeste.setUsuario(usuario);
 
-        login.setToken(token);
+            // passa a senha do usuario para codificada
+            String token = Hashing.sha256()
+                .hashString(loginTeste.getUsuario().getSenha(), StandardCharsets.UTF_8)
+                .toString(); 
 
-        Usuario usuario = _usuarioRepository.getById(1); // mudar para optional
-        login.setUsuario(usuario);
-        return _loginRepository.save(login);
-    }
+            loginTeste.setToken(token);
+            loginTeste.setAtivo(true);
+            loginTeste.setData_hora(LocalDateTime.now());
 
-    // Atualizar
-    @RequestMapping(value = "/login/{id}", method =  RequestMethod.PUT)
-    public ResponseEntity<Login> put(@PathVariable(value = "id") Integer id, @RequestBody Login newLogin)
-    {
-        Optional<Login> oldLogin = _loginRepository.findById(id);
-        if(oldLogin.isPresent()){
-            Login login = oldLogin.get();
-            login.setEmail(newLogin.getEmail());
-            login.setSenha(newLogin.getSenha());
-            login.setToken(newLogin.getToken());
-            _loginRepository.save(login);
-            return new ResponseEntity<Login>(login, HttpStatus.OK);
-        }
-        else
+            _loginRepository.save(loginTeste);
+
+            return new ResponseEntity<Login>(loginTeste, HttpStatus.OK);
+        
+        } else{
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }        
+              
     }
 
-    // Deletar
-    @RequestMapping(value = "/login/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<Object> delete(@PathVariable(value = "id") Integer id)
+    /**
+     * TODO: criar o endpoint put para atualizar
+     * 
+     * @param id
+     * @return 
+     */
+
+
+    // Deletar 
+    @RequestMapping(value = "/loginTeste/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<Object> Delete(@PathVariable(value = "id") Integer id)
     {
-        Optional<Login> login = _loginRepository.findById(id);
-        if(login.isPresent()){
-            _loginRepository.delete(login.get());
+        Optional<Login> loginTeste = _loginRepository.findById(id);
+        if(loginTeste.isPresent()){
+            _loginRepository.delete(loginTeste.get());
             return new ResponseEntity<>(HttpStatus.OK);
         }
         else
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
-    
+
 }
+
+    
+
+    
+    
+    
+
