+++ src/main/java/br/edu/ifpb/acomidadobebeservice/model/Ingrediente.java
@@ -1,16 +1,13 @@
 package br.edu.ifpb.acomidadobebeservice.model;
 
-import java.util.List;
-
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.persistence.Column;
-import javax.persistence.ManyToMany;
-import javax.persistence.JoinTable;
 import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -20,26 +17,18 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "tb_ingrediente")
+@Table(name = "tb_ingrediente_teste")
 public class Ingrediente {
+    
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_ingrediente")
     private Integer id;
     @Column(name = "nome_ingrediente")
     private String nome;
     
-
-    /**
-     * TODO: mudar o relacionamento para OneToOne
-     */
-    
-    @ManyToMany
-    @JoinTable(
-        name = "ingrediente_grupo",
-        joinColumns = @JoinColumn(name = "id_ingrediente"),
-        inverseJoinColumns = @JoinColumn(name = "id_grupo")
-    )
-    private List<Grupo> grupos;
+    @OneToOne
+    @JoinColumn(name = "id_grupo_alimentar")
+    private GrupoAlimentar grupoAlimentar;
 
 }
