+++ src/main/java/br/edu/ifpb/acomidadobebeservice/model/Item.java
@@ -20,8 +20,9 @@ import lombok.NoArgsConstructor;
 @Entity
 @Table(name = "tb_item")
 public class Item {
+    
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_item")
     private Integer id;
