+++ src/main/java/br/edu/ifpb/acomidadobebeservice/model/Login.java
@@ -1,11 +1,15 @@
 package br.edu.ifpb.acomidadobebeservice.model;
 
+import java.time.LocalDateTime;
+
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
-import javax.persistence.Column;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -15,29 +19,33 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name="tb_login")
+@Table(name = "tb_login_teste")
 public class Login {
+
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_login")
     private Integer id;
-    @Column(name = "email")
-    private String email;
-    @Column(name = "senha")
-    private String senha;
-    @Column(name = "token")
-    private String token;
-    
-    @Column(name = "id_usuario")
+
+    @OneToOne
+    @JoinColumn(name = "id_usuario")
     private Usuario usuario;
 
-    // ele vem apenas o email e senha v
+    // sha256 
+    @Column(name = "token")
+    private String token;
 
-    //
+    // do momento em que foi criado
+    @Column(name = "data_hora")
+    private LocalDateTime data_hora;
 
-    // tem que gerar o token com sha256 no controller v
+    @Column(name = "isAtivo")
+    private boolean isAtivo;
 
-    // persistir no banco e retornar o objeto completo
-
-    // criar endpoint login/token com retorno de usuario, ele pesquisa usuario pelo token / login/token/{token}
+    /**
+     * TODO 
+     * 
+     * criar endpoint login/token com retorno de usuario, ele pesquisa usuario pelo token / login/token/{token}
+     */
+    
 }
