+++ src/main/java/br/edu/ifpb/acomidadobebeservice/model/Nutricionista.java
@@ -21,7 +21,7 @@ import lombok.NoArgsConstructor;
 public class Nutricionista extends Usuario {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_nutricionista")
     private Integer id;
     @Column(name = "crn_nutricionista")
