+++ src/main/java/br/edu/ifpb/acomidadobebeservice/model/Usuario.java
@@ -22,6 +22,7 @@ import lombok.NoArgsConstructor;
 @Inheritance(strategy = InheritanceType.JOINED)
 @Table(name = "tb_usuario")
 public class Usuario implements Serializable{
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
@@ -36,4 +37,5 @@ public class Usuario implements Serializable{
     private String senha;
     @Column(name = "nascimento_usuario")
     private Date nascimento;
+    
 }
