+++ src/main/java/br/edu/ifpb/acomidadobebeservice/repository/Ingrediente1Repository.java
@@ -3,9 +3,9 @@ package br.edu.ifpb.acomidadobebeservice.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.edu.ifpb.acomidadobebeservice.model.IngredienteTeste;
+import br.edu.ifpb.acomidadobebeservice.model.Ingrediente1;
 
 @Repository
-public interface IngredienteTesteRepository extends JpaRepository<IngredienteTeste, Integer>{
+public interface Ingrediente1Repository extends JpaRepository<Ingrediente1, Integer> {
     
 }
