+++ src/main/java/br/edu/ifpb/acomidadobebeservice/repository/LoginRepository.java
@@ -1,15 +1,24 @@
 package br.edu.ifpb.acomidadobebeservice.repository;
 
-import java.util.Optional;
+//import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+//import org.springframework.data.jpa.repository.Query;
 
 import br.edu.ifpb.acomidadobebeservice.model.Login;
 
-@Repository
-public interface LoginRepository extends JpaRepository<Login, Integer> {
+public interface LoginRepository extends JpaRepository<Login, Integer>{
 
-    Optional<Login> findByToken(Integer token);
+    /**
+     * 
+     * @param token
+     * @return usuario
+     * 
+     * TODO: retorna vazio. olhar se o tipo de retorno está certo e o formato da consulta e se é nessa classe
+     *
+     */
+    
+    //@Query("FROM LoginTeste l WHERE l.usuario.token = :token ")
+    //List<LoginTeste> findUsuarioByLoginTesteToken(String token);
     
 }
