+++ src/main/resources/application.properties
@@ -1,25 +1,10 @@
-# Data Source properties
-#spring.datasource.url= jdbc:postgresql://localhost:5431/acomidadobebe
-#spring.datasource.url= jdbc:postgresql://db:5432/acomidadobebe
-#spring.datasource.username= postgres
-#spring.datasource.password= 12345
-
-#spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
-
-# Hibernate ddl auto (create, create-drop, validate, update)
-#spring.jpa.hibernate.ddl-auto= update
-
-
-#spring.datasource.driverClassName=org.postgresql.Driver
-
 # Set context base path
 # server.servlet.context-path=/api
 
 # JPA properties 
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation= true
 
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/acomidadobebe
+spring.datasource.url=jdbc:postgresql://localhost:5432/acomidadobebeteste
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.username=postgres
 spring.datasource.password=12345
